Fix SSH config parsing bug when Hostname is specified
When the et client makes a connection, it parses the available SSH
configuration files and looks for applicable Host stanzas. A Host
stanza is considered applicable when its associated pattern matches
the hostname passed to et on the command line.

In some cases multiple Host stanzas are applicable. However, if the
config file(s) contains multiple applicable Host stanzas, and one of
the earlier ones specifies the Hostname attribute, later applicable
Host stanzas will be ignored. This appears to be a bug.

The problem appears to be that the field used to store the Hostname
value is the same field used for matching against the Host stanza's
patterns. This means that once the Hostname is encountered and
stored, it clobbers the value used for matching against Host
patterns.

This commit modifies the config parsing code to use a new, immutable
variable to represent the hostname passed on the command line. It
fixes the bug described above.