Run EternalTerminal server as a Type=simple systemd service
Using Type=forking is discouraged, since systemd works best when the
processes are just running in foreground.

Change to Type=simple and drop the --daemon argument.

We don't need a PIDFile in this service unit either.

Also stopped seeing this message on the journal, upon restart of the
service (not sure exactly why this was happening):

  systemd[1]: et.service: Supervising process 12345 which is not our child. We'll most likely not notice when it exits.

Tested by making these changes to a host and ensuring I was still able
to connect to the et server.