Fix reconnects on FreeBSD
With etserver v6.1.9, attempts to reconnect cause the etserver process
to crash due to an unexpected EGAIN in UnixSocketHandler::close.

Suppressing this by ignoring EAGAIN allows etserver to keep running
but the reconnect still fails since the call to close causes the
select in TerminalServer::runTerminal to complete with EBADF and
an attempt to read from terminalFd which returns EAGAIN (since it
is non-blocking) and the terminal server session ends.

Handling EGAIN in both places allows the reconnect to succeed and
re-join the terminal server session reliably.