Add GitHub devcontainer config
This enables using things like GitHub Codespaces, which pre-configures
a container with the proper dependencies so that it can build.

I choose Ubuntu 21.04 as the base image, and needed to make one small
modification over the README instructions:

* Installed `libcurl4-openssl-dev` instead of `libcurl-dev`

The files originate from the C++ devcontainer template, with minor
modifications to install the dependency packages in `Dockerfile`.