Support for ipv6 addresses (with or without port specified)
Currently et does not support passing an ipv6 address as the host
positional argument is assumed to be an ipv4 address or hostname with a
single colon with the port after it and the rest is ignored, eg:

```
$ et 2620:10d:c083:1418:14ac:6bbc:4737:56ee                                                                                                                                     15:19:52  10.07.22  100% █
Could not reach the ET server: 2620:10
```

This change will support ipv4 address or hostname with port specified or
ipv6 address with or without port specified.  Currently, if a port is
specified in the positional arg, it overwrites what was specified
(explicitly or by default) with the -p,--port option. I think this is
backwards and may offer a PR to invert this logic.