Rename et --prefix option to --terminal-path.
The prefix option to the terminal client is misleading as it isn't a
prefix (or search path) and is treated as an absolute path to the
etterminal executable.  This renames it as such and removes the
superfluous positional arg in the command string building for the ssh
command (i.e. /usr/local/bin/etterminal etterminal -v 0)