Support for ipv6 addresses (with or without port specified) (#536)
* Support for ipv6 addresses (with or without port specified)

Currently et does not support passing an ipv6 address as the host
positional argument is assumed to be an ipv4 address or hostname with a
single colon with the port after it and the rest is ignored, eg:

```
$ et 2620:10d:c083:1418:14ac:6bbc:4737:56ee                                                                                                                                     15:19:52  10.07.22  100% █
Could not reach the ET server: 2620:10
```

This change will support ipv4 address or hostname with port specified or
ipv6 address with or without port specified.  Currently, if a port is
specified in the positional arg, it overwrites what was specified
(explicitly or by default) with the -p,--port option. I think this is
backwards and may offer a PR to invert this logic.

* etterminal option/description cleanup plus other housecleaning.

No-op changes:
* TerminalMain: change binary name/description
* TerminalMain: get rid of Daemon log file (does nothing now)
* SubprocessToString: reorganize forking logic and remove irrelevant
  comments and cleanup trailing whitespace
* PsuedoUserTerminal: breaks in forking switch statement

* Rename et --prefix option to --terminal-path.

The prefix option to the terminal client is misleading as it isn't a
prefix (or search path) and is treated as an absolute path to the
etterminal executable.  This renames it as such and removes the
superfluous positional arg in the command string building for the ssh
command (i.e. /usr/local/bin/etterminal etterminal -v 0)