Explicitly set verbosity from cxxopts value.
While investigating #541, I ran `et` with `--verbose=9` but realized
that the logs created by `et` and `etterminal` were not verbose.  It
turns out that the default argv parsing that easyloggingpp does via
START_EASYLOGGINGPP(argc, argv) does not mesh well with the cxxopts
style options so instead of relying on the default handling of argv/argc
in ELPP, let's just explicitly set the verbose level from the cxxopts
option argument.  Note that a value greater than 9 defaults to 9 in ELPP
so no handling in et/etterminal/etserver is required.

Examples of the discrepancies:

* `et --verbose=5` (cxxopts 5, ELPP 0)
** ELPP doesn't check for long opt style arg values for `--verbose` (I
have PR out for this)
* `et --verbose 5` (cxxopts 5, ELPP 9)
** ELPP doesn't detect the arg for --verbose so sets to max verbosity
* `et -v 5`        (cxxopts 5, ELPP 9)
** ELPP doesn't detect the arg for --verbose so sets to max verbosity
* `et --v=5`       (cxxopts exception, ELPP 5)
** ELPP would accept this, but cxxopts fails since we don't have a
'long' opt of `v`