Parse telemetry ini config as boolean and make telemetry opt-in.
In issue #504, it was brought to attention that while the command line
arg for enabling telemetry was boolean, the ini config in et.cfg needed
to be an int to change the value.  SimpleIni has a GetBoolValue function
that treats the strings "0", "off", "no", "false" as boolean false and
"1", "on", "true" as boolean true.

Also, this will make telemetry opt-in as opposed to opt-out as has been
requested by multiple users.  The supplied et.cfg file will still have
telemetry on.