external/cxxopts: update submodule
Needed in order to fix cross build with GCC 12.

Note that the submodule is updated to revision
c74846a891b3cc3bfa992d588b1295f528d43039 which is equal to the tagged
version 3.0.0.

This decision was made due to the changes in master of cxxopts breaking
EternalTerminal (such as removing cxxopts::OptionException from the
code).